backend/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.0
pymupdf>=1.24.0
easyocr>=1.7.0
pillow>=10.0.0
numpy>=1.26.0
torch>=2.2.0
